requirements.txt
pypdf
langchain_google_genai
chromadb
langchain
python-dotenv
langchain-text-splitters